Move Atomix server to Atomix agent.
