+++ agent/pom.xml
@@ -23,8 +23,8 @@
   </parent>
 
   <packaging>bundle</packaging>
-  <artifactId>atomix-server</artifactId>
-  <name>Atomix Server</name>
+  <artifactId>atomix-agent</artifactId>
+  <name>Atomix Agent</name>
 
   <dependencies>
     <dependency>
@@ -73,7 +73,7 @@
         <configuration>
           <createDependencyReducedPom>true</createDependencyReducedPom>
           <shadedArtifactAttached>true</shadedArtifactAttached>
-          <finalName>atomix-server</finalName>
+          <finalName>atomix-agent</finalName>
           <filters>
             <filter>
               <artifact>*:*</artifact>
@@ -97,7 +97,7 @@
               <transformers>
                 <transformer implementation="org.apache.maven.plugins.shade.resource.ServicesResourceTransformer" />
                 <transformer implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
-                  <mainClass>io.atomix.server.AtomixServer</mainClass>
+                  <mainClass>io.atomix.agent.AtomixAgent</mainClass>
                 </transformer>
               </transformers>
             </configuration>
